+++ src/crypto/encoding/base16.c
@@ -20,7 +20,7 @@ int libp2p_crypto_encoding_base16_encode(const unsigned char* incoming, size_t i
 	memset(results, 0, *results_length);
 
 	*results_length = 0;
-	for(int i = 0; i < incoming_length; i++) {
+	for(int i = 0; i < (int)incoming_length; i++) {
 		unsigned char buf[3];
 		sprintf((char*)buf, "%02x", incoming[i]);
 		results[i * 2] = buf[0];
@@ -58,7 +58,7 @@ int libp2p_crypto_encoding_base16_decode(const unsigned char* incoming, size_t i
 
 	char* pos = (char*)incoming;
 
-	for(int i = 0; i < incoming_length / 2; i++) {
+	for(int i = 0; i < (int)incoming_length / 2; i++) {
 		sscanf(pos, "%2hhx", &results[i]);
 		pos += 2;
 	}
